tests: Add auth db relation to integration tests
